requirements.txt
asgiref==3.3.1
certifi==2020.11.8
chardet==3.0.4
Django==3.1.3
facebook-sdk==3.1.0
gunicorn==20.0.4
idna==2.10
pytz==2020.4
requests==2.24.0
sqlparse==0.4.1
urllib3==1.25.11
whitenoise==5.2.0
